protocol.jtag_svf: Ensure run_count value passed to pulse_tck is integer

+++ software/glasgow/applet/jtag_svf/__init__.py
@@ -133,7 +133,7 @@ class JTAGSVFInterface(SVFEventHandler):
         if run_clock != "TCK":
             raise GlasgowAppletError("RUNTEST clock %s is not supported" % run_count)
         if run_count is None or min_time is not None and run_count / self._frequency < min_time:
-            run_count = self._frequency * min_time
+            run_count = int(self._frequency * min_time)
         if max_time is not None and run_count / self._frequency > max_time:
             self._logger.warning("RUNTEST exceeds maximum time: %d cycles (%.3f s) > %.3f s"
                                  % (run_count, run_count / self._frequency, max_time))
